Remove argent from CI
