+++ .github/workflows/e2e.yml
@@ -50,7 +50,7 @@ jobs:
 
       - name: Run tests on iOS
         working-directory: ${{ env.WORKING_DIRECTORY }}
-        run: yarn e2e:ios
+        run: yarn e2e:ios-ci
 
   android-e2e:
     if: github.repository == 'software-mansion/react-native-gesture-handler'
@@ -101,4 +101,4 @@ jobs:
           disable-animations: true
           avd-name: Pixel_9
           working-directory: ${{ env.WORKING_DIRECTORY }}
-          script: adb devices && yarn e2e:android
+          script: adb devices && yarn e2e:android-ci
