+++ apps/expo-example/e2e/__tests__/all_wo_argent.test.ts
@@ -1,15 +1,11 @@
 import { flingTests } from '../suites/gestures/fling';
 import { longPressTests } from '../suites/gestures/long_press';
 import { panTests } from '../suites/gestures/pan';
-import { pinchTests } from '../suites/gestures/pinch';
-import { rotationTests } from '../suites/gestures/rotation';
 import { tapTests } from '../suites/gestures/tap';
 import { competingTests } from '../suites/relations/competing';
 
 tapTests();
-pinchTests();
 flingTests();
 longPressTests();
-rotationTests();
 panTests();
 competingTests();
