+++ apps/expo-example/e2e/__tests__/gestures/fling.test.ts
@@ -1,3 +1,3 @@
-import { flingTests } from '../../suites/fling';
+import { flingTests } from '../../suites/gestures/fling';
 
 flingTests();
