+++ apps/expo-example/e2e/__tests__/gestures/long_press.test.ts
@@ -1,3 +1,3 @@
-import { longPressTests } from '../../suites/long_press';
+import { longPressTests } from '../../suites/gestures/long_press';
 
 longPressTests();
