+++ apps/expo-example/e2e/__tests__/gestures/rotation.test.ts
@@ -1,3 +1,3 @@
-import { rotationTests } from '../../suites/rotation';
+import { rotationTests } from '../../suites/gestures/rotation';
 
 rotationTests();
