+++ apps/expo-example/e2e/__tests__/gestures/tap.test.ts
@@ -1,3 +1,3 @@
-import { tapTests } from '../../suites/tap';
+import { tapTests } from '../../suites/gestures/tap';
 
 tapTests();
