+++ apps/expo-example/e2e/__tests__/relations/competing.test.ts
@@ -1,0 +1,3 @@
+import { competingTests } from '../../suites/relations/competing';
+
+competingTests();
