+++ apps/expo-example/e2e/suites/gestures/fling.ts
@@ -2,7 +2,7 @@
 import { beforeAll, describe } from '@jest/globals';
 import { by, element, expect } from 'detox';
 
-import { CB, navigateTo, TestScreens } from '../utils';
+import { CB, navigateTo, TestScreens } from '../../utils';
 
 export function flingTests() {
   describe('test fling gesture', () => {
