+++ apps/expo-example/e2e/suites/gestures/long_press.ts
@@ -2,7 +2,7 @@
 import { beforeAll, describe } from '@jest/globals';
 import { by, element, expect } from 'detox';
 
-import { CB, navigateTo, TestScreens } from '../utils';
+import { CB, navigateTo, TestScreens } from '../../utils';
 
 export function longPressTests() {
   describe('test long press gesture', () => {
