+++ apps/expo-example/e2e/suites/gestures/pinch.ts
@@ -4,32 +4,29 @@ import { spawn } from 'node:child_process';
 import { beforeAll, describe } from '@jest/globals';
 import { by, element, expect } from 'detox';
 
-import { CB, navigateTo, TestScreens } from '../utils';
+import { CB, navigateTo, TestScreens } from '../../utils';
 
-type RotationArgs = {
+type PinchArgs = {
   centerX: string;
   centerY: string;
-  radius: string;
-  startAngle: string;
-  endAngle: string;
+  startDistance: string;
+  endDistance: string;
 };
 
-function argentRotate(udid: string, ra: RotationArgs): Promise<boolean> {
+function argentPinch(udid: string, pa: PinchArgs): Promise<boolean> {
   const child = spawn('argent', [
     'run',
-    'gesture-rotate',
+    'gesture-pinch',
     '--udid',
     udid,
     '--centerX',
-    ra.centerX,
+    pa.centerX,
     '--centerY',
-    ra.centerY,
-    '--startAngle',
-    ra.startAngle,
-    '--endAngle',
-    ra.endAngle,
-    '--radius',
-    ra.radius,
+    pa.centerY,
+    '--startDistance',
+    pa.startDistance,
+    '--endDistance',
+    pa.endDistance,
   ]);
   return new Promise((resolve, reject) => {
     child.on('exit', (code) => {
@@ -45,37 +42,36 @@ function argentRotate(udid: string, ra: RotationArgs): Promise<boolean> {
   });
 }
 
-export function rotationTests() {
-  describe('test rotation gesture', () => {
+export function pinchTests() {
+  describe('test pinch gesture', () => {
     beforeAll(async () => {
-      await navigateTo(TestScreens.Rotation);
+      await navigateTo(TestScreens.Pinch);
     });
 
-    const gestureBox = element(by.id('rotation-box'));
+    const gestureBox = element(by.id('pinch-box'));
     const stateIndicator = element(by.id('state-indicator'));
     const extractButton = element(by.id('extract-button'));
 
-    test('Should register rotation gesture on rotation', async () => {
+    test('Should register pinch gesture on pinch', async () => {
       const udid = device.id;
 
-      await argentRotate(udid, {
+      await argentPinch(udid, {
         centerX: '0.5',
         centerY: '0.55',
-        startAngle: '0',
-        endAngle: '90',
-        radius: '0.05',
+        startDistance: '0.2',
+        endDistance: '0.6',
       });
 
       await extractButton.tap();
       await expect(stateIndicator).toHaveText(
-        `{Rotation: ${CB.B}${CB.A}${CB.U}${CB.D}${CB.F}}`
+        `{Pinch: ${CB.B}${CB.A}${CB.U}${CB.D}${CB.F}}`
       );
     });
 
-    test('Shouldn`t register a rotation gesture on tap', async () => {
+    test('Shouldn`t register a pinch gesture on tap', async () => {
       await gestureBox.tap();
       await extractButton.tap();
-      await expect(stateIndicator).toHaveText(`{Rotation: }`);
+      await expect(stateIndicator).toHaveText(`{Pinch: }`);
     });
   });
 }
